Extract interface from ExpressionExperimentPlatformSwitchService
This is needed to write unit tests for DataUpdater.
